+++ adlfs/tests/test_fetch_range.py
@@ -1,9 +1,5 @@
 from adlfs import AzureBlobFileSystem
-
-URL = "http://127.0.0.1:10000"
-ACCOUNT_NAME = "devstoreaccount1"
-KEY = "Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw=="  # NOQA
-CONN_STR = f"DefaultEndpointsProtocol=http;AccountName={ACCOUNT_NAME};AccountKey={KEY};BlobEndpoint={URL}/{ACCOUNT_NAME};"  # NOQA
+from adlfs.tests.constants import CONN_STR
 
 
 def test_fetch_entire_blob(storage):